Refaktoroi: piilota oppijatiedot jo fasadissa;
Tällä tavalla on epätodennäköisempää, että piilotus unohdetaan tehdä, kun tehdään uusia toiminnallisuuksia.
