+++ src/main/scala/fi/oph/koski/omattiedot/OmatTiedotEditorModel.scala
@@ -1,19 +1,14 @@
 package fi.oph.koski.omattiedot
 
-import java.time.LocalDate
-
 import fi.oph.koski.config.KoskiApplication
 import fi.oph.koski.editor.OppijaEditorModel.oppilaitoksenOpiskeluoikeudetOrdering
 import fi.oph.koski.editor._
 import fi.oph.koski.http.HttpStatus
-import fi.oph.koski.huoltaja.{Huollettava, HuollettavatSearchResult, HuollettavienHakuOnnistui}
+import fi.oph.koski.huoltaja.Huollettava
 import fi.oph.koski.koskiuser.KoskiSpecificSession
-import fi.oph.koski.schema.PerusopetuksenOpiskeluoikeus._
 import fi.oph.koski.schema._
 import fi.oph.koski.schema.annotation.Hidden
 import fi.oph.koski.util.{Timing, WithWarnings}
-import fi.oph.scalaschema.annotation.SyntheticProperty
-import mojave._
 
 object OmatTiedotEditorModel extends Timing {
 
@@ -24,17 +19,7 @@ object OmatTiedotEditorModel extends Timing {
     toEditorModel(userOppija.getIgnoringWarnings, näytettäväOppija.getIgnoringWarnings, warnings = näytettäväOppija.warnings)
 
   def toEditorModel(userOppija: Oppija, näytettäväOppija: Oppija, warnings: Seq[HttpStatus])(implicit application: KoskiApplication, koskiSession: KoskiSpecificSession): EditorModel = timed("createModel") {
-    buildModel(buildView(piilotetuillaTiedoilla(userOppija), piilotetuillaTiedoilla(näytettäväOppija), warnings))
-  }
-
-  def piilotetuillaTiedoilla(oppija: Oppija)(implicit koskiSession: KoskiSpecificSession): Oppija = {
-    val piilota = piilotaArvosanatKeskeneräisistäSuorituksista _ andThen
-      piilotaSensitiivisetHenkilötiedot andThen
-      piilotaKeskeneräisetPerusopetuksenPäättötodistukset andThen
-      piilotaTietojaSuoritusjaosta andThen
-      piilotaLaajuuksia
-
-    piilota(oppija)
+    buildModel(buildView(userOppija, näytettäväOppija, warnings))
   }
 
   def opiskeluoikeudetOppilaitoksittain(oppija: Oppija): List[OppilaitoksenOpiskeluoikeudet] = {
@@ -57,90 +42,6 @@ object OmatTiedotEditorModel extends Timing {
 
   private def buildModel(obj: AnyRef)(implicit application: KoskiApplication, koskiSession: KoskiSpecificSession): EditorModel = {
     EditorModelBuilder.buildModel(EditorSchema.deserializationContext, obj, editable = false)(koskiSession, application.koodistoViitePalvelu, application.koskiLocalizationRepository)
-  }
-
-  private def piilotaArvosanatKeskeneräisistäSuorituksista(oppija: Oppija) = {
-    val keskeneräisetTaiLiianÄskettäinVahvistetut = traversal[Suoritus].filter { s =>
-      s.vahvistus.isEmpty || !s.vahvistus.exists { v => v.päivä.plusDays(4).isBefore(LocalDate.now())}
-    }.compose(päätasonSuorituksetTraversal)
-    val piilotettavatOppiaineidenArvioinnit = (oppimääränArvioinnitTraversal ++ vuosiluokanArvioinnitTraversal ++ oppiaineenOppimääränArvioinnitTraversal).compose(keskeneräisetTaiLiianÄskettäinVahvistetut)
-    val piilotettavaKäyttäytymisenArviointi = käyttäytymisenArviointiTraversal.compose(keskeneräisetTaiLiianÄskettäinVahvistetut)
-
-    List(piilotettavaKäyttäytymisenArviointi, piilotettavatOppiaineidenArvioinnit).foldLeft(oppija) { (oppija, traversal) =>
-      traversal.set(oppija)(None)
-    }
-  }
-
-  private def piilotaLaajuuksia(oppija: Oppija)= {
-    val  keskenTaiVahvistettuEnnenLeikkuriPäivää = traversal[Suoritus].filter { suoritus =>
-      (suoritus.isInstanceOf[PerusopetuksenVuosiluokanSuoritus] || suoritus.isInstanceOf[NuortenPerusopetuksenOppimääränSuoritus]) &&
-        suoritus.vahvistus.forall(_.päivä.isBefore(LocalDate.of(2020, 8, 1)))
-    }.compose(päätasonSuorituksetTraversal)
-
-    val piilotettavatLaajuudet = nuortenPerusopetuksenPakollistenOppiaineidenLaajuudetTraversal.compose(keskenTaiVahvistettuEnnenLeikkuriPäivää)
-
-    piilotettavatLaajuudet.set(oppija)(None)
-  }
-
-  private def piilotaTietojaSuoritusjaosta(oppija: Oppija)(implicit koskiSession: KoskiSpecificSession) = {
-    if (koskiSession.user.isSuoritusjakoKatsominen) {
-      piilotaLukuvuosimaksutiedot(oppija)
-    } else {
-      oppija
-    }
-  }
-
-  private def piilotaLukuvuosimaksutiedot(oppija: Oppija)(implicit koskiSession: KoskiSpecificSession) = {
-    val korjatutOpiskeluoikeudet = oppija.opiskeluoikeudet.map {
-      case oo: KorkeakoulunOpiskeluoikeus if oo.lisätiedot.nonEmpty => {
-        val korjatutLukukausiIlmottautuminen = oo.lisätiedot.get.lukukausiIlmoittautuminen.flatMap(ilmo =>
-          Some(ilmo.copy(
-            ilmoittautumisjaksot = ilmo.ilmoittautumisjaksot.map(_.copy(
-              maksetutLukuvuosimaksut = None
-            ))
-          ))
-        )
-
-        val korjatutLisätiedot = oo.lisätiedot.get.copy(
-          maksettavatLukuvuosimaksut = None,
-          lukukausiIlmoittautuminen = korjatutLukukausiIlmottautuminen
-        )
-        oo.copy(
-          lisätiedot = Some(korjatutLisätiedot)
-        )
-      }
-      case oo: Any => oo
-    }
-    oppija.copy(
-      opiskeluoikeudet = korjatutOpiskeluoikeudet
-    )
-  }
-
-  private def piilotaSensitiivisetHenkilötiedot(oppija: Oppija) = {
-    val t: Traversal[Oppija, TäydellisetHenkilötiedot] = traversal[Oppija].field[Henkilö]("henkilö").ifInstanceOf[TäydellisetHenkilötiedot]
-    t.modify(oppija)((th: TäydellisetHenkilötiedot) => th.copy(hetu = None, kansalaisuus = None, turvakielto = None))
-  }
-
-  def piilotaKeskeneräisetPerusopetuksenPäättötodistukset(oppija: Oppija): Oppija = {
-    def poistaKeskeneräisetPäättötodistukset = (suoritukset: List[PäätasonSuoritus]) => suoritukset.filter(_ match {
-      case s: PerusopetuksenOppimääränSuoritus if !s.valmis => false
-      case _ => true
-    })
-
-    def poistaOsasuoritukset = (suoritukset: List[PäätasonSuoritus]) => suoritukset.map(s =>
-      shapeless.lens[PäätasonSuoritus].field[Option[List[Suoritus]]]("osasuoritukset").set(s)(None)
-    )
-
-    shapeless.lens[Oppija].field[Seq[Opiskeluoikeus]]("opiskeluoikeudet").modify(oppija)(_.map(oo => {
-      val isKeskeneräinenPäättötodistusAinoaSuoritus = oo.suoritukset match {
-        case (s: PerusopetuksenOppimääränSuoritus) :: Nil if s.kesken => true
-        case _ => false
-      }
-
-      shapeless.lens[Opiskeluoikeus].field[List[PäätasonSuoritus]]("suoritukset").modify(oo)(
-        if (isKeskeneräinenPäättötodistusAinoaSuoritus) poistaOsasuoritukset else poistaKeskeneräisetPäättötodistukset
-      )
-    }))
   }
 }
