+++ src/main/scala/fi/oph/koski/omattiedot/OmatTiedotServletV2.scala
@@ -18,7 +18,6 @@ class OmatTiedotServletV2(implicit val application: KoskiApplication) extends Ko
     renderEither[Oppija](
       huoltajaService
         .findUserOppijaAllowEmpty(session)
-        .map(_.map(OmatTiedotEditorModel.piilotetuillaTiedoilla))
         .map(_.getIgnoringWarnings)
     )
   }
